+++ Sources/StreamChat/Models/ChannelArchivingScope.swift
@@ -1,0 +1,11 @@
+//
+// Copyright © 2024 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+
+/// The scope of the channel archiving action.
+public enum ChannelArchivingScope: String {
+    /// Channel is archived only for the currently connected user. Other channel members do not see channels as archived.
+    case me
+}